muuid fix for pascal
git-svn-id: http://svn.miranda-ng.org/main/trunk@2222 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ include/delphi/m_api.pas
@@ -134,7 +134,7 @@ type
     authorEmail:PAnsiChar;
     copyright  :PAnsiChar;
     homepage   :PAnsiChar;
-    flags      :Byte;  // right now the only flag, UNICODE_AWARE, is recognized here
+    flags      :Integer;  // right now the only flag, UNICODE_AWARE, is recognized here
     uuid       :MUUID; // plugin's unique identifier
   end;
